backend/requirements.txt
Django==5.2.4
gunicorn==23.0.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
dj-database-url==3.0.1
whitenoise==6.9.0
sqlparse==0.5.3
PyJWT==2.10.1
asgiref==3.9.1